keep class definition inside NettyIoThreadFactory
